refactor(constraints): extract common reducers code
